MOD：测试增量还是全量

+++ android/src/main/java/com/fego/android/service/ReactManager.java
@@ -106,6 +106,8 @@ public class ReactManager {
      * 用来临时记录增量还是全量
      */
     private String type = null;
+    private boolean isAll = false;
+    String rnZipName = "";
     /**
      * 用来临时记录zip包的md5值
      */
@@ -336,14 +338,23 @@ public class ReactManager {
             for (String line : lines) {
                 String[] infos = line.split("_");
                 if (infos.length > 1) {
-                    String remoteSdkVersion = infos[0];
-                    String remoteDataVersion = infos[1];
-                    String localDataVer = infos[2];
-                    if (!localDataVer.equals(localDataVersion)) {
-                        continue;
+                    String remoteSdkVersion = "";
+                    String remoteDataVersion ="";
+                    remoteSdkVersion = infos[0];
+                    remoteDataVersion = infos[1];
+                    if (infos.length == 3) {
+                        isAll = true;
+                        md5Value = infos[2];
+                        type = "1";
+                    } else {
+                        isAll = false;
+                        String localDataVer = infos[2];
+                        if (!localDataVer.equals(localDataVersion)) {
+                            continue;
+                        }
+                        type = infos[3];
+                        md5Value = infos[4];
                     }
-                    type = infos[3];
-                    md5Value = infos[4];
                     if (remoteSdkVersion.equals(SDK_VERSION)) {
                         //如果新版本字典存在,说明是已下载还没有使用的资源,如果跟线上的版本号相同也不必要下载了
                         String needUpdateVersion = ReactPreference.getInstance().getString(NEW_BUNDLE_VERSION);
@@ -379,8 +390,14 @@ public class ReactManager {
      */
     private void loadRNSource(final String remoteDataVersion) {
         ReactService service = new ReactService();
-        final String rnZipName = "rn_" + SDK_VERSION + "_" + remoteDataVersion + "_" + localDataVersion + "_" + type + ".zip";
-        String rnSourceUrl = sourceUrl + SDK_VERSION + "/" + remoteDataVersion + "/" + rnZipName;
+        String rnSourceUrl = "";
+        if (isAll) {
+            rnZipName = SDK_VERSION + "_" + remoteDataVersion;
+            rnSourceUrl = sourceUrl + "all/" + SDK_VERSION + "/" + rnZipName;
+        }else {
+            rnZipName = "rn_" + SDK_VERSION + "_" + remoteDataVersion + "_" + localDataVersion + "_" + type + ".zip";
+            rnSourceUrl = sourceUrl + "increment/" + SDK_VERSION + "/" + remoteDataVersion + "/" + rnZipName;
+        }
         bundleCall = service.downloadFile(rnSourceUrl, new Callback<ResponseBody>() {
             @Override
             public void onResponse(Call<ResponseBody> call, Response<ResponseBody> response) {
@@ -431,7 +448,6 @@ public class ReactManager {
      */
     public void unzipBundle() {
         String downloadFilePath = ReactPreference.getInstance().getString(NEW_BUNDLE_PATH);
-        String remoteDataVersion = ReactPreference.getInstance().getString(NEW_BUNDLE_VERSION);
         String rnDir = sourceDir;
         File fileRNDir = new File(rnDir);
         if (!fileRNDir.exists()) {
@@ -448,18 +464,15 @@ public class ReactManager {
                 merge(patchStr, assetsBundle, rnDir);
             }
             //c、解析assetsConfig.txt，获取到需要删除的资源文件，进而删除
-            byte[] bytes = FileUtils.readFile(rnDir + "assetsConfig.txt");
-            String configDetail = new String(bytes);
-            checkAssetconfigFile(configDetail, rnDir);
-            //d、重新加载bundle
-//            onJSBundleLoadedFromServer();
-            //e、更新字体文件
+            if (!isAll) {
+                byte[] bytes = FileUtils.readFile(rnDir + "assetsConfig.txt");
+                String configDetail = new String(bytes);
+                checkAssetconfigFile(configDetail, rnDir);
+                FileUtils.deleteFile(rnDir + "increment.jsbundle");
+                FileUtils.deleteFile(rnDir + "assetsConfig.txt");
+            }
             updateReactFonts();
-            //f、删除不需要的存储和文件
-//            ReactPreference.getInstance().save(BUNDLE_VERSION, remoteDataVersion);
             FileUtils.delete(file);
-            FileUtils.deleteFile(rnDir + "increment.jsbundle");
-            FileUtils.deleteFile(rnDir + "assetsConfig.txt");
         }
     }
